refactored so sequence is again a derivative of viewer
