+++ app/node/output/track/tracklist.cpp
@@ -24,6 +24,7 @@
 #include "node/math/math/math.h"
 #include "node/math/merge/merge.h"
 #include "node/output/viewer/viewer.h"
+#include "project/item/sequence/sequence.h"
 
 namespace olive {
