+++ app/project/item/sequence/sequence.h
@@ -21,23 +21,8 @@
 #ifndef SEQUENCE_H
 #define SEQUENCE_H
 
-#include <QUuid>
-
-#include "common/rational.h"
-#include "node/block/block.h"
-#include "node/graph.h"
-#include "node/node.h"
-#include "node/output/track/track.h"
 #include "node/output/track/tracklist.h"
-#include "node/traverser.h"
-#include "project/item/footage/footage.h"
-#include "render/audioparams.h"
-#include "render/audioplaybackcache.h"
-#include "render/framehashcache.h"
-#include "render/videoparams.h"
-#include "render/videoparams.h"
-#include "timeline/timelinecommon.h"
-#include "timeline/timelinepoints.h"
+#include "node/output/viewer/viewer.h"
 #include "timeline/timelinepoints.h"
 
 namespace olive {
@@ -45,11 +30,11 @@ namespace olive {
 /**
  * @brief The main timeline object, an graph of edited clips that forms a complete edit
  */
-class Sequence : public Node, public TimelinePoints
+class Sequence : public ViewerOutput
 {
   Q_OBJECT
 public:
-  Sequence(bool viewer_only_mode = false);
+  Sequence();
 
   virtual ~Sequence() override;
 
@@ -82,13 +67,6 @@ public:
 
   virtual QIcon icon() const override;
 
-  virtual QString duration() const override;
-  virtual QString rate() const override;
-
-  void set_default_parameters();
-
-  void set_parameters_from_footage(const QVector<Footage *> footage);
-
   const QVector<Track *> &GetTracks() const
   {
     return track_cache_;
@@ -109,67 +87,13 @@ public:
     return track_lists_.at(type);
   }
 
-  void ShiftVideoCache(const rational& from, const rational& to);
-  void ShiftAudioCache(const rational& from, const rational& to);
-  void ShiftCache(const rational& from, const rational& to);
-
-  virtual void InvalidateCache(const TimeRange& range, const QString& from, int element, qint64 job_time) override;
-
-  VideoParams video_params() const
-  {
-    return GetStandardValue(kVideoParamsInput).value<VideoParams>();
-  }
-
-  AudioParams audio_params() const
-  {
-    return GetStandardValue(kAudioParamsInput).value<AudioParams>();
-  }
-
-  void set_video_params(const VideoParams &video)
-  {
-    SetStandardValue(kVideoParamsInput, QVariant::fromValue(video));
-  }
-
-  void set_audio_params(const AudioParams &audio)
-  {
-    SetStandardValue(kAudioParamsInput, QVariant::fromValue(audio));
-  }
-
-  const rational &GetLength() const;
-
   virtual void Retranslate() override;
 
-  FrameHashCache* video_frame_cache()
-  {
-    return &video_frame_cache_;
-  }
-
-  AudioPlaybackCache* audio_playback_cache()
-  {
-    return &audio_playback_cache_;
-  }
-
-  virtual void BeginOperation() override;
-
-  virtual void EndOperation() override;
-
-  static const QString kVideoParamsInput;
-  static const QString kAudioParamsInput;
-
-  static const QString kTextureInput;
-  static const QString kSamplesInput;
   static const QString kTrackInputFormat;
-
-  static const uint64_t kVideoParamEditMask;
 
   TimelinePoints* timeline_points()
   {
     return &timeline_points_;
-  }
-
-  const QUuid& uuid() const
-  {
-    return uuid_;
   }
 
   virtual bool IsItem() const override
@@ -177,61 +101,27 @@ public:
     return true;
   }
 
-signals:
-  void TimebaseChanged(const rational&);
-
-  void LengthChanged(const rational& length);
-
-  void SizeChanged(int width, int height);
-
-  void PixelAspectChanged(const rational& pixel_aspect);
-
-  void InterlacingChanged(VideoParams::Interlacing mode);
-
-  void VideoParamsChanged();
-  void AudioParamsChanged();
-
-  void TrackAdded(Track* track);
-  void TrackRemoved(Track* track);
-
-  void TextureInputChanged();
-
-public slots:
-  void VerifyLength();
-
 protected:
-  virtual void InputConnectedEvent(const QString &input, int element, const NodeOutput &output) override;
-
-  virtual void InputDisconnectedEvent(const QString &input, int element, const NodeOutput &output) override;
-
-  virtual rational GetCustomLength(Track::Type type) const;
-
-  virtual void ShiftVideoEvent(const rational &from, const rational &to);
-
-  virtual void ShiftAudioEvent(const rational &from, const rational &to);
-
   virtual void LoadInternal(QXmlStreamReader* reader, XMLNodeData &xml_node_data, uint version, const QAtomicInt* cancelled) override;
 
   virtual void SaveInternal(QXmlStreamWriter *writer) const override;
 
-  virtual void InputValueChangedEvent(const QString& input, int element) override;
+  virtual void ShiftAudioEvent(const rational &from, const rational &to) override;
+
+  virtual void InputConnectedEvent(const QString &input, int element, const NodeOutput &output) override;
+
+  virtual void InputDisconnectedEvent(const QString &input, int element, const NodeOutput &output) override;
+
+  virtual rational GetCustomLength(Track::Type type) const override;
+
+signals:
+  void TrackAdded(Track* track);
+  void TrackRemoved(Track* track);
 
 private:
   QVector<TrackList*> track_lists_;
 
   QVector<Track*> track_cache_;
-
-  QUuid uuid_;
-
-  rational last_length_;
-
-  FrameHashCache video_frame_cache_;
-
-  AudioPlaybackCache audio_playback_cache_;
-
-  int operation_stack_;
-
-  VideoParams cached_video_params_;
 
   TimelinePoints timeline_points_;
