+++ app/task/render/render.h
@@ -35,7 +35,7 @@ class RenderTask : public Task
 {
   Q_OBJECT
 public:
-  RenderTask(Sequence* viewer, const VideoParams &vparams, const AudioParams &aparams);
+  RenderTask(ViewerOutput* viewer, const VideoParams &vparams, const AudioParams &aparams);
 
   virtual ~RenderTask() override;
 
@@ -53,7 +53,7 @@ protected:
 
   virtual void AudioDownloaded(const TimeRange& range, SampleBufferPtr samples, qint64 job_time) = 0;
 
-  Sequence* viewer() const
+  ViewerOutput* viewer() const
   {
     return viewer_;
   }
@@ -87,7 +87,7 @@ private:
 
   void StartTicket(const QByteArray &hash, QThread *watcher_thread, ColorManager *manager, const rational &time, RenderMode::Mode mode, FrameHashCache *cache, const QSize &force_size, const QMatrix4x4 &force_matrix, VideoParams::Format force_format, ColorProcessorPtr force_color_output);
 
-  Sequence* viewer_;
+  ViewerOutput* viewer_;
 
   VideoParams video_params_;
