+++ app/widget/curvewidget/curvewidget.h
@@ -67,7 +67,7 @@ protected:
 
   virtual void TimeTargetChangedEvent(Node* target) override;
 
-  virtual void ConnectedNodeChanged(Sequence* n) override;
+  virtual void ConnectedNodeChanged(ViewerOutput* n) override;
 
 private:
   void SetKeyframeButtonEnabled(bool enable);
