+++ app/widget/timebased/timebasedwidget.h
@@ -46,9 +46,11 @@ public:
 
   void ZoomOut();
 
-  Sequence* GetConnectedNode() const;
+  ViewerOutput* GetConnectedNode() const;
 
-  void ConnectViewerNode(Sequence *node);
+  TimelinePoints* GetConnectedTimelinePoints() const;
+
+  void ConnectViewerNode(ViewerOutput *node);
 
   void SetScaleAndCenterOnPlayhead(const double& scale);
 
@@ -105,21 +107,19 @@ protected:
 
   virtual void ScaleChangedEvent(const double &) override;
 
-  virtual void ConnectedNodeChanged(Sequence*){}
+  virtual void ConnectedNodeChanged(ViewerOutput*){}
 
-  virtual void ConnectNodeInternal(Sequence*){}
+  virtual void ConnectNodeInternal(ViewerOutput*){}
 
-  virtual void DisconnectNodeInternal(Sequence*){}
+  virtual void DisconnectNodeInternal(ViewerOutput*){}
 
   void SetAutoMaxScrollBar(bool e);
 
   virtual void resizeEvent(QResizeEvent *event) override;
 
-  virtual TimelinePoints* ConnectTimelinePoints();
+  virtual TimelinePoints* GetTimelinePointsToConnect();
 
   virtual Project* GetTimelinePointsProject();
-
-  TimelinePoints* GetConnectedTimelinePoints() const;
 
   void ConnectTimelineView(TimeBasedView* base, bool connect_time_change_event = true);
 
@@ -189,7 +189,7 @@ private:
 
   bool UserIsDraggingPlayhead() const;
 
-  Sequence* viewer_node_;
+  ViewerOutput* viewer_node_;
 
   TimeRuler* ruler_;
